Fixed invalidation duplication
