+++ app/gang/[id]/battle-session/[sessionId]/page.tsx
@@ -2,7 +2,8 @@ import { createClient } from '@/utils/supabase/server';
 import { redirect, notFound } from 'next/navigation';
 import { getAuthenticatedUser, signInPath } from '@/utils/auth';
 import { getBattleSessionCached } from '@/app/lib/battle-sessions/get-battle-session-data';
-import { getGangFightersList, getGangPositioning, type GangFighter } from '@/app/lib/shared/gang-data';
+import { getGangFightersList, getGangPositioning } from '@/app/lib/shared/gang-data';
+import type { GangFighter } from '@/types/gang';
 import { getCampaignTerritories } from '@/app/lib/campaigns/[id]/get-campaign-data';
 import { getScenariosCached } from '@/app/lib/reference-data';
 import { checkCampaignArbitrator } from '@/utils/user-permissions';
